fix(ci): clear clippy lints the workspace gate now enforces
Newer clippy (1.96) flags two pre-existing lints the gate added in the
previous commit surfaces: redundant `as u32` casts in status.rs tests and
an identity `1 * 49 * 40 * 1` in the micro_speech test.

+++ crates/iree-embedded/src/status.rs
@@ -122,13 +122,13 @@ mod tests {
 
     #[test]
     fn aborted_maps_to_error() {
-        let st = status_of(sys::IREE_STATUS_ABORTED as u32);
+        let st = status_of(sys::IREE_STATUS_ABORTED);
         assert_eq!(check(st).unwrap_err().code(), StatusCode::Aborted);
     }
 
     #[test]
     fn resource_exhausted_maps_to_oom() {
-        let st = status_of(sys::IREE_STATUS_RESOURCE_EXHAUSTED as u32);
+        let st = status_of(sys::IREE_STATUS_RESOURCE_EXHAUSTED);
         assert_eq!(check(st).unwrap_err().code(), StatusCode::OutOfMemory);
     }
 }
